requirements.txt
aiogram==3.22.0    # ← ОБНОВЛЕНО! Исправлены баги с Pydantic
pydantic==2.10.6   # ← ФИКС! Совместимая версия с aiogram 3.22.0
python-dotenv==1.0.1
aiosqlite==0.20.0
aiofiles==23.2.1